+++ packages/composites/augur/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @chainlink/augur-adapter
 
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies [fe8ab13c]
+- Updated dependencies [b78f8e06]
+- Updated dependencies [c93e5654]
+- Updated dependencies [ccff5d7f]
+  - @chainlink/therundown-adapter@1.1.0
+  - @chainlink/ea-bootstrap@1.3.0
+  - @chainlink/sportsdataio-adapter@1.1.5
+
 ## 1.0.6
 
 ### Patch Changes