+++ packages/sources/dxfeed-secondary/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @chainlink/dxfeed-secondary-adapter
 
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies [b78f8e06]
+- Updated dependencies [c93e5654]
+- Updated dependencies [ccff5d7f]
+  - @chainlink/ea-bootstrap@1.3.0
+  - @chainlink/ea-test-helpers@1.0.1
+  - @chainlink/dxfeed-adapter@1.1.3
+
 ## 1.1.2
 
 ### Patch Changes