+++ packages/sources/json-rpc/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @chainlink/json-rpc-adapter
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [b78f8e06]
+- Updated dependencies [c93e5654]
+- Updated dependencies [ccff5d7f]
+  - @chainlink/ea-bootstrap@1.3.0
+  - @chainlink/ea-test-helpers@1.0.1
+
 ## 1.1.1
 
 ### Patch Changes